Improve module organization by splitting app_config and renaming rule modules
Config changes:
- Split app_config.rs into focused modules: web_ui.rs, agents.rs, loader.rs
- Improved separation of concerns for WebUI, agent, and configuration loading logic
- Updated all imports to use new module structure

Rule module renaming for clarity:
- Renamed diff_monitor.rs → when.rs (DiffMonitor → When) for when condition processing
- Renamed timeout.rs → diff_timeout.rs (TimeoutMonitor → DiffTimeout) for diff_timeout rules
- Updated module references and imports throughout codebase

These changes make the codebase more intuitive by aligning module names with their actual functionality.
